% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{biVarimax}
\alias{biVarimax}
\title{biVarimax}
\usage{
biVarimax(U, V, lambda = 1, normalize = TRUE, eps = 1e-05)
}
\arguments{
\item{lambda}{hyper-parameter of balancing U and V: obj(U, V)  = lambda * obj_varimax(U) + obj_varimax(V). As lambda increases, more emphasis is put on U. Default is 1.}

\item{normalize}{logical; if true, the rows of U, V are re-scaled to unit length before rotation, and scaled back afterwards. Default is TRUE.}

\item{eps}{the tolerance for stopping: the relative change in the sum of singular values. Default is 1e-5.}

\item{V:}{phenotype - mediator matrix; each row represents a phenotype, each column represents a mediator. The number of mediators should be no larger than that of phenotypes.}

\item{U:}{genotype - mediator matrix estimator; each row represents a genotype, each column represents a mediator. The number of mediators should be no larger than that of genotypes.}
}
\description{
rotate a matrix pair (U, V) to have clear structures. Extended from varimax of a single matrix.
}
