% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlot.R
\name{clusterPlot}
\alias{clusterPlot}
\title{clusterPlot}
\usage{
clusterPlot(model, pValue = NULL, names = TRUE)
}
\arguments{
\item{g}{...}

\item{parameters:}{visualization.method, tSNE.perplexity, clustering.method, cluster.centers, pValue.k, pValue.threshold, rsnps.read, saveAddress, randomSeed, UMAP.n_neighbors, genotypeWeight, betweenWeight, phenotypeWeight}
}
\description{
clustering and visualization
}
