% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coappearance.R
\name{coappearance.list}
\alias{coappearance.list}
\title{order each type of edges decreasingly in edge weights.}
\usage{
coappearance.list(g, top = 10)
}
\arguments{
\item{g}{a weighted graph object from igraph. By default, there are three types of edges: gene-gene, gene-trait, and trait-trait. Genes are recorded as "rsXXX", and traits do not start with "rs".}

\item{top}{how many edges for display.}
}
\description{
order each type of edges decreasingly in edge weights.
}
\examples{
SR0 = readRDS(file = "/Users/zijungao/Dropbox/PathGPS/plot/source code/source data/metabolitesUMAP.rds")
coappearance.list(SR0$graph)

}
