% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{colSpDistance}
\alias{colSpDistance}
\title{colSpDistance}
\usage{
colSpDistance(X1, X2)
}
\arguments{
\item{X}{covariate matrix to predict conditional densities at; each row represents an observation vector.}
}
\description{
compute the distance between the col space of X1 and that of X2
}
