% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnEstimator.R
\name{columnEstimator}
\alias{columnEstimator}
\title{columnEstimator}
\usage{
columnEstimator(
  betas = "metabolites",
  p,
  pNull,
  pSample = NULL,
  r,
  m = 100,
  V.subtract = TRUE,
  V.rotate = TRUE,
  V.rotate.m = 1,
  U.project = TRUE,
  randomSeed = 318
)
}
\arguments{
\item{randomSeed}{}

\item{X}{covariate matrix to predict conditional densities at; each row represents an observation vector.}
}
\value{
VHat, UHat
}
\description{
estimate the row space of V (#mediators x #phenotypes) and the column space of U (#genotypes x #mediators)
}
