% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{latentMediatorExample}
\alias{latentMediatorExample}
\title{latentMediatorExample}
\usage{
latentMediatorExample(
  data = "metabolites",
  plotOnly = TRUE,
  clustering.method = "UMAP",
  envr.subtract = TRUE
)
}
\arguments{
\item{X}{...}
}
\description{
latentMediatorExample
}
