% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnEstimator.R
\name{pathGPS.oneshot}
\alias{pathGPS.oneshot}
\title{pathGPS.oneshot}
\usage{
pathGPS.oneshot(
  beta1,
  beta0,
  r,
  envr.subtract = TRUE,
  sprs.transform = TRUE,
  parameters = list()
)
}
\description{
oneshot estimation of U and V
}
