% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tSNEPlot.R
\name{tSNEPlot}
\alias{tSNEPlot}
\title{tSNEPlot}
\usage{
tSNEPlot(
  g,
  maximalSimilarity = 1,
  tSNE.perplexity = 5,
  clusteringMethod = "kmeans",
  clustering.centers = "label propagation",
  pValue.k = 150,
  pValue.threshold = NULL,
  rsnps.read = FALSE,
  saveAddress = NULL,
  randomSeed = 100
)
}
\arguments{
\item{X}{covariate matrix to predict conditional densities at; each row represents an observation vector.}
}
\description{
tSNE plot and kmeans clusters
}
