% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{UVEstimator}
\alias{UVEstimator}
\title{UVEstimator}
\usage{
UVEstimator(
  beta,
  beta2,
  r,
  subtract = TRUE,
  rotate = TRUE,
  hardThresholding = FALSE,
  sparse = FALSE,
  parameters = NULL
)
}
\arguments{
\item{parameters}{list of parameters.}

\item{beta:}{non-null marginal association matrix; each row represents a non-null genotype, each column represents a phenotype}

\item{beta2:}{null marginal association matrix; each row represents a null genotype, each column represents a phenotype}

\item{r:}{number of latent mediators/pathways.}

\item{subtract:}{logical; if true, subtraction of null marginal association matrix from non-null marginal association matrix is conducted.}

\item{rotate:}{logical; if true, rotate the estimated U, V to have clear structures.}

\item{sparse:}{logical; if true, spca from package elasticnet is used to estimate sparse U, V.}

\item{hardThresholding:}{logical; if true, estimated U, V are hard-thresholded to preserve the desired number of non-zero entries.}
}
\description{
estimate the phenotype - mediator matrix V and the genotype - mediator matrix U
}
