% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterPlot.R
\name{clusterPlot}
\alias{clusterPlot}
\title{clusterPlot}
\usage{
clusterPlot(g, maximalSimilarity = 1, parameters = list())
}
\arguments{
\item{g}{...}

\item{parameters:}{visualization.method, tSNE.perplexity, clustering.method, cluster.centers, pValue.k, pValue.threshold, rsnps.read, saveAddress, randomSeed, UMAP.n_neighbors, genotypeWeight, betweenWeight, phenotypeWeight}
}
\description{
clustering and visualization
}
