% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnEstimator.R
\name{columnEstimator}
\alias{columnEstimator}
\title{columnEstimator}
\usage{
columnEstimator(
  betas = "metabolites",
  p,
  pNull,
  pSample = NULL,
  r,
  V.subtract = TRUE,
  rotate = TRUE,
  rotate.m = 1,
  U.project = TRUE,
  UV.truncate = FALSE,
  aggregate.method = "bootstrap",
  parameters = list()
)
}
\arguments{
\item{X}{covariate matrix to predict conditional densities at; each row represents an observation vector.}

\item{aggregate.method:}{"subsample", "bootstrap"}

\item{parameters:}{list of parameters: rotate.m, propUZero, probVZero, randomSeed, similarity.truncate, n.subsample}
}
\value{
VHat, UHat
}
\description{
estimate the row space of V (#mediators x #phenotypes) and the column space of U (#genotypes x #mediators)
}
